verify/requirements.txt
flask
faiss-cpu
google-generativeai
langchain
langchain-google-genai
numpy
flask-cors
tweepy
textblob
python-dotenv